99.9% kraj

+++ app/Http/Controllers/UserController.php
@@ -13,9 +13,14 @@ class UserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        $users = User::all();
-        return view('admin_users', compact('users'));
+    {   
+        if(auth()->user()->is_admin){
+            $users = User::all();
+            return view('admin_users', compact('users'));
+        }else{
+            return back()->with('message', "You don't have privilages to do that!");
+        }
+        
     }
 
     /**
@@ -35,21 +40,27 @@ class UserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {
-        $formFields = $request->validate([
-            'name' => ['required', 'min:3'],
-            'username' => ['required'],
-            'password' => ['required'],
-            
-
-        ]);
-
-
-        $formFields['password'] = bcrypt($formFields['password']);
-
-        $user = User::create($formFields);
-
-        return redirect('/admin/index')->with('message', 'User created successfully!');
+    {   
+        
+        if(auth()->user()->is_admin){
+            $formFields = $request->validate([
+                'name' => ['required', 'min:3'],
+                'username' => ['required'],
+                'password' => ['required'],
+                
+    
+            ]);
+    
+    
+            $formFields['password'] = bcrypt($formFields['password']);
+    
+            $user = User::create($formFields);
+    
+            return redirect('/admin/index')->with('message', 'User created successfully!');
+        }else{
+            return back()->with('message', "You don't have privilages to do that!");
+        }
+        
 
     }
 
@@ -72,9 +83,15 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $user = User::whereId($id)->get();
-        $user = $user[0];
-        return view('admin_update_user', compact('user'));
+
+        if(auth()->user()->is_admin){
+            $user = User::whereId($id)->get();
+            $user = $user[0];
+            return view('admin_update_user', compact('user'));
+        }else{
+            return back()->with('message', "You don't have privilages to do that!");
+        }
+        
     }
 
     /**
@@ -85,16 +102,21 @@ class UserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
-    {
-        $data = $request->validate([
-            'name' => 'required',
-            'username' => 'required',
-            'password' => 'required',
-            'is_admin' => 'required',
-        ]);
-
-        User::whereId($id)->get()->first()->update($data);
-        return redirect('staff')->with('message', 'User updated successfully!');
+    {   
+        if(auth()->user()->is_admin){
+            $data = $request->validate([
+                'name' => 'required',
+                'username' => 'required',
+                'password' => 'required',
+                'is_admin' => 'required',
+            ]);
+    
+            User::whereId($id)->get()->first()->update($data);
+            return redirect('staff')->with('message', 'User updated successfully!');
+        }else{
+            return back()->with('message', "You don't have privilages to do that!");
+        }
+        
     }
 
     /**
@@ -104,10 +126,15 @@ class UserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Request $request)
-    {
-        User::whereId($request->delete)->delete();
+    {   
+        if(auth()->user()->is_admin){
+            User::whereId($request->delete)->delete();
 
-        return back()->with('message', 'User deleted successfully!');
+            return back()->with('message', 'User deleted successfully!');
+        }else{
+            return back()->with('message', "You don't have privilages to do that!");
+        }
+        
     }
 
     public function login(Request $request, User $user)
